requirements.txt
matplotlib
plotly
streamlit
pyspark
kafka-python
pymongo
psycopg2-binary
pandas
